2019年7月12日

+++ script/sh_link.sh
@@ -9,7 +9,6 @@ echo -n "$1" \
  | sed -e 's@#@♯@g' \
  | sed -e 's@\r@_@g' \
  | sed -e 's@\n@_@g' \
- | sed -e 's@ @_@g' \
  | sed -e 's@,@，@g' \
  | sed -e 's@+@➕@g' \
  | sed -e 's@=@↔️@g' \
@@ -22,7 +21,7 @@ echo -n "$1" \
  | sed -e "s@>@》@g" \
  | sed -e 's@;@🔚@g' \
  | sed -e 's@`@▪️@g' \
- | sed -e 's@:@📎@g' \
+ | sed -e 's@:@：@g' \
  | sed -e 's@!@❗️@g' \
  | sed -e 's@*@✳️@g' \
  | sed -e 's@?@❓@g' \
@@ -37,6 +36,8 @@ echo -n "$1" \
  | sed -e "s@'@▫️@g" \
  | sed -e 's@"@”@g'
  
+# | sed -e 's@ @_@g'
+
 }
 
 add_ss_link () {
@@ -123,6 +124,7 @@ ss_link_plugin_opts=""
 
 clear_link () {
 
+logger -t "【SS】" "服务器订阅：清空上次订阅节点配置"
 # 自定义节点配置靠前保存
 mkdir -p /tmp/ss/link
 ss_x=`nvram get rt_ssnum_x`
@@ -142,6 +144,11 @@ done
 # 删除🔗订阅连接
 cat /tmp/ss/link/daochu_1.txt | sort -u | grep -v "^$" > /tmp/ss/link/daochu_2.txt
 grep "🔗" /tmp/ss/link/daochu_2.txt | cut -d '=' -f1 | awk -F '_x' '{print $2}' | sort -u > /tmp/ss/link/daochu_3.txt
+if [ ! -s /tmp/ss/link/daochu_3.txt ] ; then
+    echo "不含订阅连接"
+    rm -rf /tmp/ss/link
+    return
+fi
 while read line
 do
     sed -Ei "/rt_ss_name_x$line=|rt_ss_port_x$line=|rt_ss_password_x$line=|rt_ss_server_x$line=|rt_ss_usage_x$line=|rt_ss_method_x$line=/d" /tmp/ss/link/daochu_2.txt
@@ -201,7 +208,7 @@ rm -rf /tmp/ss/link
 
 do_link () {
 mkdir -p /tmp/ss/link
-logger -t "【SS】" "订阅文件下载: $ssr_link_i"
+#logger -t "【SS】" "订阅文件下载: $ssr_link_i"
 rm -f /tmp/ss/link/0_link.txt
 wgetcurl.sh /tmp/ss/link/0_link.txt "$ssr_link_i" "$ssr_link_i" N
 if [ ! -s /tmp/ss/link/0_link.txt ] ; then
@@ -249,6 +256,8 @@ sed -e "s/_/\//g" -i /tmp/ss/link/ssr_link.txt
 sed -e "s/\-/\+/g" -i /tmp/ss/link/ssr_link.txt
 
 #echo > /tmp/ss/link/c_link.txt
+[ -f /www/link/link.js ] && echo -n "var ACL2List = [" > /www/link/link.js
+i_s=0
 i=`nvram get rt_ssnum_x`
 if [ -f /tmp/ss/link/ssr_link.txt ] ; then
 	awk  'BEGIN{FS="\n";}  {cmd=sprintf("echo -n %s|base64 -d", $1);  system(cmd); print "";}' /tmp/ss/link/ssr_link.txt > /tmp/ss/link/ssr_link2.txt
@@ -258,13 +267,26 @@ if [ -f /tmp/ss/link/ssr_link.txt ] ; then
 		add_0
 		add_ssr_link "$line"
 		#echo  $ss_link_name $ss_link_server $ss_link_port $ss_link_password $ss_link_method $ss_link_obfs $ss_link_protocol >> /tmp/ss/link/c_link.txt
-		eval "nvram set rt_ss_name_x$i=\"🔗$ss_link_name\""
-		eval "nvram set rt_ss_port_x$i=$ss_link_port"
-		eval "nvram set rt_ss_password_x$i=\"$ss_link_password\""
-		eval "nvram set rt_ss_server_x$i=$ss_link_server"
-		eval "nvram set rt_ss_usage_x$i=\"-o $ss_link_obfs -O $ss_link_protocol $ss_link_obfsparam $ss_link_protoparam\""
-		eval "nvram set rt_ss_method_x$i=$ss_link_method"
-		i=$(( i + 1 ))
+		if [ ! -f /www/link/link.js ] ; then
+			eval "nvram set rt_ss_name_x$i=\"🔗$ss_link_name\""
+			eval "nvram set rt_ss_port_x$i=$ss_link_port"
+			eval "nvram set rt_ss_password_x$i=\"$ss_link_password\""
+			eval "nvram set rt_ss_server_x$i=$ss_link_server"
+			eval "nvram set rt_ss_usage_x$i=\"-o $ss_link_obfs -O $ss_link_protocol $ss_link_obfsparam $ss_link_protoparam\""
+			eval "nvram set rt_ss_method_x$i=$ss_link_method"
+			i=$(( i + 1 ))
+		else
+			[ $i_s -gt 0 ] && echo -n ', ' >> /www/link/link.js
+			echo -n '["🔗'"$ss_link_name"'", ' >> /www/link/link.js
+			echo -n '"'"$ss_link_server"'", ' >> /www/link/link.js
+			echo -n '"'"$ss_link_port"'", ' >> /www/link/link.js
+			echo -n '"'"$ss_link_password"'", ' >> /www/link/link.js
+			echo -n '"'"$ss_link_method"'", ' >> /www/link/link.js
+			ping_link
+			echo -n '"'"-o $ss_link_obfs -O $ss_link_protocol $ss_link_obfsparam $ss_link_protoparam"'", ' >> /www/link/link.js
+			echo -n '""]' >> /www/link/link.js
+			i_s=$(( i_s + 1 ))
+		fi
 	fi
 	done < /tmp/ss/link/ssr_link2.txt
 fi
@@ -278,25 +300,63 @@ if [ -f /tmp/ss/link/ss_link.txt ] ; then
 		add_0
 		add_ss_link "$line"
 		#echo  $ss_link_name $ss_link_server $ss_link_port $ss_link_password $ss_link_method $ss_link_obfs $ss_link_protocol >> /tmp/ss/link/c_link.txt
-		eval "nvram set rt_ss_name_x$i=\"🔗$ss_link_name\""
-		eval "nvram set rt_ss_port_x$i=$ss_link_port"
-		eval "nvram set rt_ss_password_x$i=\"$ss_link_password\""
-		eval "nvram set rt_ss_server_x$i=$ss_link_server"
-		eval "nvram set rt_ss_method_x$i=$ss_link_method"
-		eval "nvram set rt_ss_usage_x$i=\"$ss_link_plugin_opts\""
-		i=$(( i + 1 ))
+		if [ ! -f /www/link/link.js ] ; then
+			eval "nvram set rt_ss_name_x$i=\"🔗$ss_link_name\""
+			eval "nvram set rt_ss_port_x$i=$ss_link_port"
+			eval "nvram set rt_ss_password_x$i=\"$ss_link_password\""
+			eval "nvram set rt_ss_server_x$i=$ss_link_server"
+			eval "nvram set rt_ss_method_x$i=$ss_link_method"
+			eval "nvram set rt_ss_usage_x$i=\"$ss_link_plugin_opts\""
+			i=$(( i + 1 ))
+		else
+			[ $i_s -gt 0 ] && echo -n ', ' >> /www/link/link.js
+			echo -n '["🔗'"$ss_link_name"'", ' >> /www/link/link.js
+			echo -n '"'"$ss_link_server"'", ' >> /www/link/link.js
+			echo -n '"'"$ss_link_port"'", ' >> /www/link/link.js
+			echo -n '"'"$ss_link_password"'", ' >> /www/link/link.js
+			echo -n '"'"$ss_link_method"'", ' >> /www/link/link.js
+			ping_link
+			echo -n '"'"$ss_link_plugin_opts"'", ' >> /www/link/link.js
+			echo -n '""]' >> /www/link/link.js
+			i_s=$(( i_s + 1 ))
+		fi
 	fi
 	done < /tmp/ss/link/ss_link2.txt
 fi
-
+[ -f /www/link/link.js ] && echo -n ']' >> /www/link/link.js
+if [ ! -f /www/link/link.js ] ; then
 # 保存有效节点数量
 rt_ssnum_x=`nvram get rt_ssnum_x`
 [ -z $rt_ssnum_x ] && rt_ssnum_x=0 && nvram set rt_ssnum_x=0
 [ $rt_ssnum_x -lt $i ] && nvram set rt_ssnum_x=$i
 rm -rf /tmp/ss/link
 nvram commit
-
+fi
 }
+
+
+ping_link () {
+
+ping_text=`ping -4 $ss_link_server -c 1 -w 2 -q`
+ping_time=`echo $ping_text | awk -F '/' '{print $4}'| awk -F '.' '{print $1}'`
+ping_loss=`echo $ping_text | awk -F ', ' '{print $3}' | awk '{print $1}'`
+if [ ! -z "$ping_time" ] ; then
+	[ $ping_time -le 250 ] && echo -n '"btn-success", ' >> /www/link/link.js
+	[ $ping_time -gt 250 ] && echo -n '"btn-warning", ' >> /www/link/link.js
+	[ $ping_time -gt 500 ] && echo -n '"btn-danger", ' >> /www/link/link.js
+	echo "🔗$ss_link_name：$ping_time ms 丢包率：$ping_loss"
+	#logger -t "【🔗$ss_link_name】" "$ping_time ms"
+	echo -n '"'"$ping_time ms"'", ' >> /www/link/link.js
+ else
+	echo -n '"btn-danger", ' >> /www/link/link.js
+	echo "🔗$ss_link_name：失效1"
+	logger -t "【🔗$ss_link_name】" "失效"
+	echo -n '"失效", ' >> /www/link/link.js
+fi
+}
+
+
+start_link () {
 
 rt_ssnum_x=$(nvram get rt_ssnum_x)
 [ -z $rt_ssnum_x ] && rt_ssnum_x="0"
@@ -308,6 +368,7 @@ rt_ssnum_x_tmp="`nvram get rt_ssnum_x_tmp`"
 if [ "$rt_ssnum_x_tmp"x = "del"x ] ; then
 	shlinksh=$$
 	eval $(ps -w | grep "sh_link.sh" | grep -v grep | grep -v "$shlinksh" | awk '{print "kill -9 "$1";";}')
+	echo -n '' > /www/link/link.js
 	clear_link
 	nvram set rt_ssnum_x_tmp=0
 	nvram commit
@@ -324,11 +385,11 @@ if [ "$A_restart" != "$B_restart" ] ; then
 	nvram set ss_link_status=$B_restart
 	if [ -z "$ssr_link" ] ; then
 		cru.sh d ss_link_update
-		logger -t "【SS】" "停止 SSR 服务器订阅"
+		logger -t "【SS】" "停止 SS 服务器订阅"
 		return
 	else
 		cru.sh a ss_link_update "12 */3 * * * $scriptfilepath uplink &" &
-		logger -t "【SS】" "启动 SSR 服务器订阅: $ssr_link"
+		logger -t "【SS】" "启动 SS 服务器订阅，添加计划任务 (Crontab)，每三小时更新"
 	fi
 fi
 if [ -z "$ssr_link" ] ; then
@@ -337,9 +398,12 @@ fi
 shlinksh=$$
 eval $(ps -w | grep "sh_link.sh" | grep -v grep | grep -v "$shlinksh" | awk '{print "kill -9 "$1";";}')
 
-logger -t "【SS】" "服务器订阅：清空上次订阅节点配置"
-# 清空上次订阅节点配置
-clear_link
+if [ ! -f /www/link/link.js ] ; then
+	# 清空上次订阅节点配置
+	clear_link
+else
+	logger -t "【SS】" "服务器订阅：开始更新"
+fi
 ssr_link="$(echo "$ssr_link" | sed 's@   @ @g' | sed 's@^ @@g' | sed 's@ $@@g' )"
 ssr_link_i=""
 if [ ! -z "$(echo "$ssr_link" | awk -F ' ' '{print $2}')" ] ; then
@@ -352,4 +416,40 @@ else
 	ssr_link_i="$ssr_link"
 	do_link
 fi
-logger -t "【SS】" "服务器订阅：完成"
+logger -t "【SS】" "服务器订阅：更新完成"
+}
+
+check_link () {
+mkdir -p /etc/storage/link
+touch /etc/storage/link/link.js
+if [ -f /www/link/link.js ] && [ ! -s /www/link/link.js ] ; then
+# 使用 /etc/storage/link 保存订阅节点
+logger -t "【SS】" "服务器订阅：由原来 NVRAM 保存节点配置，转为使用 /etc/storage/link 保存订阅节点"
+# 清空上次订阅节点配置
+clear_link
+nvram commit
+fi
+}
+
+case $ACTION in
+stop)
+	shlinksh=$$
+	eval $(ps -w | grep "sh_link.sh" | grep -v grep | grep -v "$shlinksh" | awk '{print "kill -9 "$1";";}')
+	;;
+start)
+	start_link
+	;;
+start_nvram)
+	rm -f /etc/storage/link/link.js
+	start_link
+	;;
+check)
+	check_link
+	;;
+*)
+	check_link
+	start_link
+	;;
+esac
+
+
